feat(new-design): create backend
